+++ accel-derive/Cargo.toml
@@ -18,7 +18,7 @@ proc-macro = true
 quote = "0.5"
 
 [dependencies.nvptx]
-version = "0.1"
+version = "0.2.1"
 
 [dependencies.syn]
 version = "0.13"
